requirements.txt
pyttsx3
SpeechRecognition
openai
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
numpy
pyaudio